// src/app/models/user.model.ts
export class User{
  "num": number;
  "firstName": string;
  "lastName": string;
  "email": string;
  "phone": number;
  "weight": number;
  "height": number;
  "bmi": number;
  "bmiResult": string;
  "gender": string;
  "requireTrainer": string;
  "package": string;
  "importantList": string[];
  "haveGymBefore": string;
  "date": string;
}